build-sys: requires gstreamer 1.9 or above
The recent work on spice-gtk with Gstreamer for video stream is
targeting hardware video acceleration. One of our main targets is
intel-based hardware and for that the VAAPI elements from
gstreamer-vaapi are necessary.

The gstreamer-vaapi project is maintained and released as part of
GStreamer since 1.8 [0] and the auto detection for VAAPI elements
based on client's hardware since 1.9 [1].

This patch removes checks for old GStreamer versions and request the
1.9.0 which is new enough for VAAPI. Note that we don't require the
elements itself to be present on build time.

Also, at this moment:
- Fedora 26 has 1.12
- Fedora 25 has 1.10
- Fedora 24 has 1.8 (EOL in August 8th)
- Debian 9 has 1.10
- RHEL 7 has 1.10

[0] See release notes from 1.8.0 (NEWS) 1abf889dddc75b4e4 (gstreamer)

[1] IRC chat on #gstreamer with Víctor M. Jáques
   toso | ceyusa: hey, is there a way to verify if we can hw
          decode using vaapi? some sort of api that translate
          what vainfo and validate with gstreamer-vaapi maybe?
   toso | s/what vainfo/the output of vainfo/
 ceyusa | toso: using gstreamer-vaapi 1.9, only the available
          decoder entries are registered
 ceyusa | so, using a gst-inspect-1.0 you'll see only the
          available decoders
   toso | ceyusa: and the available decoders takes in
          consideration my actual hw then?
 ceyusa | toso: yes
   toso | ceyusa: awesome

Signed-off-by: Victor Toso <victortoso@redhat.com>
Acked-by: Marc-André Lureau <marcandre.lureau@redhat.com>