session: introduce "video-sync-on-audio-latency" property
The multi media time (mmtime) is the relative clock used to sync video
and audio decoding. The mmtime is initially set on SPICE_MSG_MAIN_INIT
and readjusted on every SPICE_MSG_MAIN_MULTI_MEDIA_TIME message.

Basically, spice-server as provider of video and audio stream content
is in charge of giving us the clock reference to make audio and video
be played synchronized.

What is currently happening is that, the client is adjusting the
mmtime based on audio backend's latency input. Although this has been
working well,  a more straight forward solution is to inform
spice-server about audio's latency in order to adjust mmtime in the
source.

Considering that the video might not be encoded in the host, having
the client changing the mmtime can increase significantly the video
decoding latency.

This patch does not change how things are working now but introduce a
new property, SpiceSession::video-sync-on-audio-latency that should
help, when disabled, identify synchronization's bug.

This patch also includes a FIXME as we want to inform spice-server
about any latency. This message does not exist yet.

Signed-off-by: Victor Toso <victortoso@redhat.com>