Use libva to check video hardware accel capabilities
Libva is an implementation for VA-API.

This can be used to automatically send to the server the
preferred video codecs as the client would prefer streams
with video codecs that can be decoded with gpu support.

We can also use the profiles to detect and set upper limit for
video streams quality.
e.g: Don't start UHD video stream if client's hardware don't
support it.

This patch makes usage of libva in spice-session and exposes this
information to all available channel-displays with the internal
function spice_session_get_hw_accel_video_codecs()

Signed-off-by: Victor Toso <victortoso@redhat.com>