channel-display: don't debug latency for each frame
Becomes quite hard to find meaning on something that is printed every
time. Only print latency value if it is a new min/max or if average
latency is 10% bigger/lower then usual.

Not aiming to perfect statistics in latency here, just to avoid too
verbose logging. Removing latency debug isn't cool as we could infer
issues with streaming based on it.

Signed-off-by: Victor Toso <victortoso@redhat.com>